Measure the duration of commands.
For each command the user types in, the duration of the command in
milliseconds is now stored in `LAST_CMD_DURATION` and available to the
prompt. Also stored are `LAST_CMD`, `LAST_CMD_START` and `LAST_CMD_END`.

The prompt has a new part which shows the time in milliseconds the last
second took to run if it was longer than a second.